+++ packages/zowe-explorer/__tests__/__unit__/trees/dataset/DatasetActions.unit.test.ts
@@ -15,7 +15,6 @@ import {
     Gui,
     imperative,
     Validation,
-    Types,
     ProfilesCache,
     ZoweExplorerApiType,
     ZoweScheme,
@@ -3679,6 +3678,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
         let withProgressSpy: jest.SpyInstance;
         let showMessageSpy: jest.SpyInstance;
         let showInputBoxSpy: jest.SpyInstance;
+        let showQuickPickSpy: jest.SpyInstance;
         let errorMessageSpy: jest.SpyInstance;
         let tableBuilderOptionsSpy: jest.SpyInstance;
         let tableBuilderTitleSpy: jest.SpyInstance;
@@ -3687,6 +3687,15 @@ describe("Dataset Actions Unit Tests - function search", () => {
         let tableBuilderColumnsSpy: jest.SpyInstance;
         let tableBuilderAddRowActionSpy: jest.SpyInstance;
         let tableBuilderBuildSpy: jest.SpyInstance;
+
+        function quickPickPositiveExpect() {
+            expect(showQuickPickSpy).toHaveBeenCalledWith([vscode.l10n.t("Case Insensitive"), vscode.l10n.t("Case Sensitive")], {
+                title: vscode.l10n.t("Search Case Sensitivity"),
+                placeHolder: vscode.l10n.t("Select search case sensitivity"),
+                ignoreFocusOut: true,
+                canPickMany: false,
+            });
+        }
 
         beforeAll(() => {
             tableViewProviderSetTableViewMock = jest.fn();
@@ -3700,6 +3709,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
             showMessageSpy = jest.spyOn(Gui, "showMessage").mockImplementation();
             errorMessageSpy = jest.spyOn(Gui, "errorMessage").mockImplementation();
             showInputBoxSpy = jest.spyOn(Gui, "showInputBox");
+            showQuickPickSpy = jest.spyOn(Gui, "showQuickPick");
             tableBuilderOptionsSpy = jest.spyOn(TableBuilder.prototype, "options").mockReturnValue(TableBuilder.prototype);
             tableBuilderTitleSpy = jest.spyOn(TableBuilder.prototype, "title").mockReturnValue(TableBuilder.prototype);
             tableBuilderIsViewSpy = jest.spyOn(TableBuilder.prototype, "isView").mockReturnValue(TableBuilder.prototype);
@@ -3731,6 +3741,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
 
             expect(errorMessageSpy).toHaveBeenCalledWith("No search pattern applied. Search for a pattern and try again.");
             expect(showMessageSpy).not.toHaveBeenCalled();
+            expect(showQuickPickSpy).not.toHaveBeenCalled();
             expect(performSearchSpy).not.toHaveBeenCalled();
             expect(getSearchMatchesSpy).not.toHaveBeenCalled();
             expect(tableViewProviderSpy).not.toHaveBeenCalled();
@@ -3751,7 +3762,30 @@ describe("Dataset Actions Unit Tests - function search", () => {
 
             expect(errorMessageSpy).not.toHaveBeenCalled();
             expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
-            expect(showMessageSpy).toHaveBeenCalledWith("Operation Cancelled");
+            expect(showQuickPickSpy).not.toHaveBeenCalled();
+            expect(performSearchSpy).not.toHaveBeenCalled();
+            expect(getSearchMatchesSpy).not.toHaveBeenCalled();
+            expect(tableViewProviderSpy).not.toHaveBeenCalled();
+            expect(openSearchAtLocationSpy).not.toHaveBeenCalled();
+            expect(withProgressSpy).not.toHaveBeenCalled();
+            expect(tableBuilderTitleSpy).not.toHaveBeenCalled();
+        });
+
+        it("should fail to perform a search if the user does not specify search case sensitivity", async () => {
+            const profile = createIProfile();
+            const node = createDatasetSessionNode(createISession(), profile);
+            node.pattern = "FAKE.*.DS";
+            const context = { context: "fake" } as any;
+            const searchString = "test";
+
+            showInputBoxSpy.mockResolvedValue(searchString);
+            showQuickPickSpy.mockResolvedValue(undefined);
+
+            await DatasetActions.search(context, node);
+
+            expect(errorMessageSpy).not.toHaveBeenCalled();
+            expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
+            quickPickPositiveExpect();
             expect(performSearchSpy).not.toHaveBeenCalled();
             expect(getSearchMatchesSpy).not.toHaveBeenCalled();
             expect(tableViewProviderSpy).not.toHaveBeenCalled();
@@ -3774,6 +3808,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
             const myProgress = { test: "test" };
 
             showInputBoxSpy.mockResolvedValue(searchString);
+            showQuickPickSpy.mockResolvedValue(vscode.l10n.t("Case Insensitive"));
             withProgressSpy.mockImplementation((opts: any, fn: any) => {
                 return fn(myProgress, tokenCancellation);
             });
@@ -3783,6 +3818,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
 
             expect(errorMessageSpy).not.toHaveBeenCalled();
             expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
+            quickPickPositiveExpect();
             expect(showMessageSpy).not.toHaveBeenCalled();
             expect(performSearchSpy).toHaveBeenCalledTimes(1);
             expect(getSearchMatchesSpy).not.toHaveBeenCalled();
@@ -3837,6 +3873,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
             const myProgress = { test: "test" };
 
             showInputBoxSpy.mockResolvedValue(searchString);
+            showQuickPickSpy.mockResolvedValue(vscode.l10n.t("Case Insensitive"));
             withProgressSpy.mockImplementation((opts: any, fn: any) => {
                 return fn(myProgress, tokenCancellation);
             });
@@ -3847,6 +3884,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
 
             expect(errorMessageSpy).not.toHaveBeenCalled();
             expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
+            quickPickPositiveExpect();
             expect(showMessageSpy).not.toHaveBeenCalled();
 
             expect(withProgressSpy.mock.calls[0][0]).toEqual({
@@ -3858,6 +3896,12 @@ describe("Dataset Actions Unit Tests - function search", () => {
             expect(getSearchMatchesSpy).toHaveBeenCalledWith(node, expectedResponse, true, searchString);
 
             expect(performSearchSpy).toHaveBeenCalledTimes(1);
+            expect(performSearchSpy).toHaveBeenCalledWith(myProgress, tokenCancellation, {
+                node,
+                pattern: node.pattern,
+                searchString,
+                caseSensitive: false,
+            });
             expect(openSearchAtLocationSpy).not.toHaveBeenCalled();
 
             expect(tableBuilderTitleSpy).toHaveBeenCalledWith('Search Results for "test"');
@@ -3975,6 +4019,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
             const myProgress = { test: "test" };
 
             showInputBoxSpy.mockResolvedValue(searchString);
+            showQuickPickSpy.mockResolvedValue(vscode.l10n.t("Case Insensitive"));
             withProgressSpy.mockImplementation((opts: any, fn: any) => {
                 return fn(myProgress, tokenCancellation);
             });
@@ -3985,6 +4030,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
 
             expect(errorMessageSpy).not.toHaveBeenCalled();
             expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
+            quickPickPositiveExpect();
             expect(showMessageSpy).not.toHaveBeenCalled();
 
             expect(withProgressSpy.mock.calls[0][0]).toEqual({
@@ -3996,6 +4042,12 @@ describe("Dataset Actions Unit Tests - function search", () => {
             expect(getSearchMatchesSpy).toHaveBeenCalledWith(pdsNode, expectedResponse, false, searchString);
 
             expect(performSearchSpy).toHaveBeenCalledTimes(1);
+            expect(performSearchSpy).toHaveBeenCalledWith(myProgress, tokenCancellation, {
+                node: pdsNode,
+                pattern: pdsNode.label,
+                searchString,
+                caseSensitive: false,
+            });
             expect(openSearchAtLocationSpy).not.toHaveBeenCalled();
 
             expect(tableBuilderTitleSpy).toHaveBeenCalledWith('Search Results for "test"');
@@ -4089,6 +4141,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
             const myProgress = { test: "test" };
 
             showInputBoxSpy.mockResolvedValue(searchString);
+            showQuickPickSpy.mockResolvedValue(vscode.l10n.t("Case Insensitive"));
             withProgressSpy.mockImplementation((opts: any, fn: any) => {
                 return fn(myProgress, tokenCancellation);
             });
@@ -4099,6 +4152,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
 
             expect(errorMessageSpy).not.toHaveBeenCalled();
             expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
+            quickPickPositiveExpect();
             expect(showMessageSpy).not.toHaveBeenCalled();
 
             expect(withProgressSpy.mock.calls[0][0]).toEqual({
@@ -4110,6 +4164,12 @@ describe("Dataset Actions Unit Tests - function search", () => {
             expect(getSearchMatchesSpy).toHaveBeenCalledWith(node, expectedResponse, true, searchString);
 
             expect(performSearchSpy).toHaveBeenCalledTimes(1);
+            expect(performSearchSpy).toHaveBeenCalledWith(myProgress, tokenCancellation, {
+                node,
+                pattern: node.label,
+                searchString,
+                caseSensitive: false,
+            });
             expect(openSearchAtLocationSpy).not.toHaveBeenCalled();
 
             expect(tableBuilderTitleSpy).toHaveBeenCalledWith('Search Results for "test"');
@@ -4229,6 +4289,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
             const myProgress = { test: "test" };
 
             showInputBoxSpy.mockResolvedValue(searchString);
+            showQuickPickSpy.mockResolvedValue(vscode.l10n.t("Case Insensitive"));
             withProgressSpy.mockImplementation((opts: any, fn: any) => {
                 return fn(myProgress, tokenCancellation);
             });
@@ -4239,6 +4300,7 @@ describe("Dataset Actions Unit Tests - function search", () => {
 
             expect(errorMessageSpy).not.toHaveBeenCalled();
             expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
+            quickPickPositiveExpect();
             expect(showMessageSpy).not.toHaveBeenCalled();
 
             expect(withProgressSpy.mock.calls[0][0]).toEqual({
@@ -4250,6 +4312,131 @@ describe("Dataset Actions Unit Tests - function search", () => {
             expect(getSearchMatchesSpy).toHaveBeenCalledWith(pdsNode, expectedResponse, false, searchString);
 
             expect(performSearchSpy).toHaveBeenCalledTimes(1);
+            expect(performSearchSpy).toHaveBeenCalledWith(myProgress, tokenCancellation, {
+                node: pdsNode,
+                pattern: pdsNode.label,
+                searchString,
+                caseSensitive: false,
+            });
+            expect(openSearchAtLocationSpy).not.toHaveBeenCalled();
+
+            expect(tableBuilderTitleSpy).toHaveBeenCalledWith('Search Results for "test"');
+            expect(tableBuilderOptionsSpy).toHaveBeenCalledWith({
+                autoSizeStrategy: { type: "fitCellContents" },
+                pagination: true,
+                rowSelection: "multiple",
+                selectEverything: true,
+                suppressRowClickSelection: true,
+            });
+            expect(tableBuilderIsViewSpy).toHaveBeenCalledTimes(1);
+            expect(tableBuilderRowsSpy).toHaveBeenCalledWith(...expectedMatches);
+            expect(tableBuilderColumnsSpy).toHaveBeenCalledWith(
+                ...[
+                    {
+                        field: "name",
+                        headerName: vscode.l10n.t("Data Set Name"),
+                        filter: true,
+                        initialSort: "asc",
+                    } as Table.ColumnOpts,
+                    {
+                        field: "position",
+                        headerName: vscode.l10n.t("Position"),
+                        filter: false,
+                    },
+                    {
+                        field: "contents",
+                        headerName: vscode.l10n.t("Contents"),
+                        filter: true,
+                    },
+                    {
+                        field: "actions",
+                        hide: true,
+                    },
+                ]
+            );
+            expect(tableBuilderAddRowActionSpy).toHaveBeenCalledWith("all", {
+                title: vscode.l10n.t("Open"),
+                command: "open",
+                callback: {
+                    fn: (DatasetActions as any).openSearchAtLocation,
+                    typ: "multi-row",
+                },
+                type: "secondary",
+            });
+            expect(tableBuilderBuildSpy).toHaveBeenCalledTimes(1);
+            expect(tableViewProviderSpy).toHaveBeenCalledTimes(1);
+            expect(tableViewProviderSetTableViewMock).toHaveBeenCalledTimes(1);
+        });
+        it("should attempt to perform the search with case sensitivity", async () => {
+            const profile = createIProfile();
+            const node = createDatasetSessionNode(createISession(), profile);
+            node.pattern = "FAKE.*.DS";
+            const context = { context: "fake" } as any;
+            const searchString = "test";
+
+            const expectedResponse = {
+                success: true,
+                apiResponse: [
+                    {
+                        dsn: "FAKE.DATA.SET.DS",
+                        member: undefined,
+                        matchList: [
+                            {
+                                line: 1,
+                                column: 1,
+                                contents: "test",
+                            },
+                        ],
+                    },
+                ],
+            };
+            const expectedMatches = [
+                {
+                    name: "FAKE.DATA.SET.DS",
+                    line: 1,
+                    column: 1,
+                    position: "1:1",
+                    contents: "test",
+                    uri: "/test/FAKE.DATA.SET.DS",
+                    searchString,
+                },
+            ];
+            const tokenCancellation: vscode.CancellationToken = {
+                isCancellationRequested: false,
+                onCancellationRequested: jest.fn(),
+            };
+            const myProgress = { test: "test" };
+
+            showInputBoxSpy.mockResolvedValue(searchString);
+            showQuickPickSpy.mockResolvedValue(vscode.l10n.t("Case Sensitive"));
+            withProgressSpy.mockImplementation((opts: any, fn: any) => {
+                return fn(myProgress, tokenCancellation);
+            });
+            getSearchMatchesSpy.mockReturnValue(expectedMatches);
+            performSearchSpy.mockResolvedValue(expectedResponse);
+
+            await DatasetActions.search(context, node);
+
+            expect(errorMessageSpy).not.toHaveBeenCalled();
+            expect(showInputBoxSpy).toHaveBeenCalledWith({ prompt: "Enter the text to search for." });
+            quickPickPositiveExpect();
+            expect(showMessageSpy).not.toHaveBeenCalled();
+
+            expect(withProgressSpy.mock.calls[0][0]).toEqual({
+                location: vscode.ProgressLocation.Notification,
+                title: 'Searching for "test"',
+                cancellable: true,
+            });
+
+            expect(getSearchMatchesSpy).toHaveBeenCalledWith(node, expectedResponse, true, searchString);
+
+            expect(performSearchSpy).toHaveBeenCalledTimes(1);
+            expect(performSearchSpy).toHaveBeenCalledWith(myProgress, tokenCancellation, {
+                node,
+                pattern: node.pattern,
+                searchString,
+                caseSensitive: true,
+            });
             expect(openSearchAtLocationSpy).not.toHaveBeenCalled();
 
             expect(tableBuilderTitleSpy).toHaveBeenCalledWith('Search Results for "test"');